backend/services/news-feeds-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.25.2
pandas==2.1.4
structlog==23.2.0
